fix: improved dupe detection logic
previously, it was only considering dupes
in the context of the period. this was meant
to also reward updating. however, it didn't
take into account the edge case in which a
"ping pong" occurs between mainnet and gnosis.

considering dupes only in the period allows
users to farm and deplete most of the rewards
by resubmitting the same address over and
over again. so, it would have been a vulnerability
if left in its original state.